requirements.txt
numpy-1.24.3
paho-mqtt-1.6.1
keyboard-0.13.5
path-16.6.0
matplotlib~=3.7.1
keyboard~=0.13.5
Flask-2.3.2
mutagen-1.46
pychromecast==7.5.1
protobuf-3.20.1
waitress-2.1.2
pyttsx3-2.90
netifaces-0.11.0
paho-mqtt~=1.6.1
mutagen~=1.46.0
sounddevice~=0.4.6
numpy~=1.24.3
scipy~=1.10.1
pyttsx3~=2.90
Flask~=2.3.2
waitress~=2.1.2